1.0.1

+++ 06_reports/00.reports.compile_tpcds.sql
@@ -2,6 +2,6 @@ DROP SCHEMA IF EXISTS reports CASCADE;
 CREATE SCHEMA reports;
 
 CREATE EXTERNAL WEB TABLE reports.compile_tpcds
-(id int, description varchar, tuples int, duration time) 
+(id int, description varchar, tuples bigint, duration time) 
 EXECUTE :EXECUTE ON MASTER
 FORMAT 'TEXT' (DELIMITER '|');
