+++ 06_reports/02.reports.init.sql
@@ -1,4 +1,4 @@
 CREATE EXTERNAL WEB TABLE reports.init
-(id int, description varchar, tuples int, duration time) 
+(id int, description varchar, tuples bigint, duration time) 
 EXECUTE :EXECUTE ON MASTER
 FORMAT 'TEXT' (DELIMITER '|');
