+++ 06_reports/04.reports.load.sql
@@ -1,5 +1,5 @@
 CREATE EXTERNAL WEB TABLE reports.load
-(id int, description varchar, tuples int, duration time) 
+(id int, description varchar, tuples bigint, duration time) 
 EXECUTE :EXECUTE ON MASTER
 FORMAT 'TEXT' (DELIMITER '|');
