+++ 06_reports/05.reports.sql.sql
@@ -1,5 +1,5 @@
 CREATE EXTERNAL WEB TABLE reports.sql
-(id int, description varchar, tuples int, duration time) 
+(id int, description varchar, tuples bigint, duration time) 
 EXECUTE :EXECUTE ON MASTER
 FORMAT 'TEXT' (DELIMITER '|');
